Perusahaan_service/service_python/requirements.txt
# requirements.txt
flask
mysql-connector-python